Remove RN bottom tabs
