+++ app.json
@@ -38,7 +38,6 @@
       ],
       "expo-router",
       "expo-font",
-      "react-native-bottom-tabs",
       "expo-web-browser"
     ],
     "scheme": "skiamultiplatform",
