+++ layouts/TabLayout.tsx
@@ -1,37 +1,51 @@
+import { Tabs } from 'expo-router';
 import React from 'react';
-import { withLayoutContext } from 'expo-router';
-import { createNativeBottomTabNavigator } from '@bottom-tabs/react-navigation';
-
+import { Platform, Pressable } from 'react-native';
+import { BottomTabBarButtonProps } from '@react-navigation/bottom-tabs';
 import { Colors } from '@/constants/Colors';
 import { useColorScheme } from '@/hooks/useColorScheme';
 import { useTextStyles } from '@/hooks/useTextStyles';
 
-import TabLayoutJS from './TabLayout.web';
-import { Platform } from 'react-native';
-export const Tabs = withLayoutContext(
-  createNativeBottomTabNavigator().Navigator,
-);
-
 export default function TabLayout() {
-  if (Platform.OS === 'android') {
-    return <TabLayoutJS />;
-  }
   const colorScheme = useColorScheme();
-  const colors = Colors[colorScheme ?? 'light'];
-
   const textStyles = useTextStyles();
+
+  const tabBarButton = (props: BottomTabBarButtonProps) => {
+    const style: any = props.style ?? {};
+    return (
+      <Pressable
+        {...props}
+        style={({ pressed, focused }) => [
+          style,
+          {
+            opacity: pressed || focused ? 0.6 : 1.0,
+          },
+        ]}
+      />
+    );
+  };
 
   return (
     <Tabs
-      tabBarActiveTintColor={colors.tabIconSelected}
-      tabBarInactiveTintColor={colors.tabIconDefault}
-      rippleColor={colors.tint}
-      labeled={true}
+      screenOptions={{
+        tabBarActiveTintColor: Colors[colorScheme ?? 'light'].tint,
+        tabBarActiveBackgroundColor: Colors[colorScheme ?? 'light'].background,
+        tabBarStyle: {
+          width: '100%',
+        },
+        tabBarPosition: 'top',
+        tabBarIconStyle: {
+          height: textStyles.title.lineHeight,
+          width: 0,
+        },
+        headerShown: false,
+      }}
     >
       <Tabs.Screen
         name="index"
         options={{
           title: 'Home',
+          tabBarButton,
           tabBarLabelStyle: textStyles.default,
           tabBarIcon: () => null,
         }}
@@ -40,6 +54,7 @@ export default function TabLayout() {
         name="breathe"
         options={{
           title: 'Breathe',
+          tabBarButton,
           tabBarLabelStyle: textStyles.default,
           tabBarIcon: () => null,
         }}
@@ -48,6 +63,7 @@ export default function TabLayout() {
         name="glass"
         options={{
           title: 'Glassmorphism',
+          tabBarButton,
           tabBarLabelStyle: textStyles.default,
           tabBarIcon: () => null,
         }}
@@ -56,26 +72,48 @@ export default function TabLayout() {
         name="reanimated"
         options={{
           title: 'Reanimated',
+          tabBarButton,
           tabBarLabelStyle: textStyles.default,
           tabBarIcon: () => null,
         }}
       />
-      <Tabs.Screen
-        name="barchart"
-        options={{
-          title: 'Bar Chart',
-          tabBarLabelStyle: textStyles.default,
-          tabBarIcon: () => null,
-        }}
-      />
-      <Tabs.Screen
-        name="linechart"
-        options={{
-          title: 'Line Graph',
-          tabBarLabelStyle: textStyles.default,
-          tabBarIcon: () => null,
-        }}
-      />
+      {/* Victory-native hidden on web */}
+      {Platform.OS !== 'web' ? (
+        <Tabs.Screen
+          name="barchart"
+          options={{
+            title: 'Bar Chart',
+            tabBarButton,
+            tabBarLabelStyle: textStyles.default,
+            tabBarIcon: () => null,
+          }}
+        />
+      ) : (
+        <Tabs.Screen
+          name="barchart"
+          options={{
+            href: null,
+          }}
+        />
+      )}
+      {Platform.OS !== 'web' ? (
+        <Tabs.Screen
+          name="linechart"
+          options={{
+            title: 'Line Chart',
+            tabBarButton,
+            tabBarLabelStyle: textStyles.default,
+            tabBarIcon: () => null,
+          }}
+        />
+      ) : (
+        <Tabs.Screen
+          name="linechart"
+          options={{
+            href: null,
+          }}
+        />
+      )}
     </Tabs>
   );
 }
