+++ package.json
@@ -21,7 +21,6 @@
     "preset": "jest-expo"
   },
   "dependencies": {
-    "@bottom-tabs/react-navigation": "^0.10.0",
     "@expo/vector-icons": "^14.1.0",
     "@shopify/react-native-skia": "2.2.3",
     "expo": "^54.0.0-preview.1",
@@ -40,7 +39,6 @@
     "react": "19.1.0",
     "react-dom": "19.1.0",
     "react-native": "npm:react-native-tvos@0.81.0-0",
-    "react-native-bottom-tabs": "^0.10.0",
     "react-native-gesture-handler": "~2.28.0",
     "react-native-reanimated": "~4.0.1",
     "react-native-safe-area-context": "5.4.0",
